fix: Fix minor bug when importing utility functions from test file, instead just create new file with utility functions and import from there
